requirements.txt
flask
openai
python-dotenv
pyodbc
langchain 
langgraph
langchain-community
langchain-ollama
tabulate
langchain_core
vanna[chromadb,ollama]